working signals, signals.c created
