+++ Makefile
@@ -6,7 +6,7 @@
 #    By: asyed <asyed@student.42singapore.sg>       +#+  +:+       +#+         #
 #                                                 +#+#+#+#+#+   +#+            #
 #    Created: 2024/05/24 16:11:35 by ting              #+#    #+#              #
-#    Updated: 2024/06/19 18:39:20 by asyed            ###   ########.fr        #
+#    Updated: 2024/06/20 20:39:11 by asyed            ###   ########.fr        #
 #                                                                              #
 # **************************************************************************** #
 
@@ -25,7 +25,8 @@ SRCS = minishell.c builtins_env_node_utils.c builtin_env.c\
 		builtin_export.c builtin_unset.c builtin_cd_pwd.c\
 		parsing.c lexer.c lexer_utils.c remove_quotes.c\
 		expand_env_var.c parsing_utils.c\
-		free.c print_error_msg.c init_status_envdup.c
+		free.c print_error_msg.c init_status_envdup.c\
+		signals.c
 
 OBJS = $(addprefix $(SRCSPATH), $(SRCS:.c=.o))
