+++ includes/minishell.h
@@ -6,7 +6,7 @@
 /*   By: asyed <asyed@student.42singapore.sg>       +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/05/24 16:12:23 by ting              #+#    #+#             */
-/*   Updated: 2024/06/20 11:36:42 by asyed            ###   ########.fr       */
+/*   Updated: 2024/06/20 20:38:42 by asyed            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -70,6 +70,14 @@ typedef struct s_cmd
 //init_status_envdup.c
 t_ms_state	*init_status();
 t_env	**init_envdup(t_ms_state *status, char **env);
+
+//--------------------SIGNALS----------------------
+
+//signals.c
+void	sigexit_handler(t_cmd **cmds, t_env **env_dup, t_ms_state *status);
+void	sigignore_handler();
+void	sigquit_handler();
+int		signal_handlers_setup(void);
 
 //--------------------BUILT-INS--------------------
